+++ components/41-organisms/footer/footer.config.js
@@ -109,7 +109,8 @@ module.exports = {
           ]
         }
       }
-    }, {
+    },
+    {
       name: 'de',
       preview: '@preview-without-padding',
       context: {