+++ CHANGELOG.md
@@ -7,6 +7,15 @@ All notable changes to this style guide are documented here.
 ### Added 
 
 * Added checklist spot icon.
+* Added GentInfo language variants.
+
+### Updated
+
+* **Updated the footer template** to replace Digipolis by our new name: 'District09'.
+
+### Fixed
+
+* Fixed spacing for inline icon list.
 
 ## [4.0.2]